apps/edge-player/requirements.txt
python-mpv==1.0.8
pydantic==2.9.2
structlog==24.4.0

